updated Person controller, successful

+++ Project2Solution/Project.Client/Controllers/PersonController.cs
@@ -17,11 +17,19 @@ namespace Project.Client.Controllers
     {
         private readonly Lib.IRepository repository;
 
+        IEnumerable<Person> plist = new List<Person>();
+
         public PersonController(Lib.IRepository repository)
         {
 
             this.repository = repository;
         }
+
+        /*public PersonController(Lib.IRepository repository, IEnumerable<Person> plist)
+        {
+            this.repository = repository;
+            this.plist = plist;
+        }*/
         // GET api/values
         [HttpGet]
         [ProducesResponseType(typeof(Person), StatusCodes.Status200OK)]
@@ -30,7 +38,6 @@ namespace Project.Client.Controllers
         {
 
             IEnumerable<Person> persons = await Task.Run(() => Mapper.Map(repository.GetPersons()));
-
             return persons;
 
         }
@@ -48,9 +55,11 @@ namespace Project.Client.Controllers
 
         // POST api/values
         [HttpPost]
+        [ProducesResponseType(typeof(Person), StatusCodes.Status201Created)]
+        [Produces("application/json")]
         public async Task<IActionResult> Post([FromBody] Person person)
         {
-            if (ModelState.IsValid) return BadRequest();
+            if (!ModelState.IsValid) return BadRequest();
 
             int rowAffected = await Task.Run(() =>
             repository.CreatePerson(Mapper.Map(person)));
@@ -62,8 +71,12 @@ namespace Project.Client.Controllers
 
         //PUT api/values/5
         [HttpPut]
+        [ProducesResponseType(typeof(Person), StatusCodes.Status202Accepted)]
+        [Produces("application/json")]
         public async Task<IActionResult> Put([FromBody] Person person)
         {
+            if (!ModelState.IsValid) return BadRequest();
+
             int rowAffected = await Task.Run(() =>
             repository.UpdatePerson(Mapper.Map(person)));
 
@@ -74,13 +87,18 @@ namespace Project.Client.Controllers
 
         // DELETE api/values/5
         [HttpDelete("{id}")]
+        [ProducesResponseType(typeof(Person), StatusCodes.Status200OK)]
+        [Produces("application/json")]
         public async Task<IActionResult> Delete([FromRoute] int id)
         {
+            if (!ModelState.IsValid) return BadRequest();
+            
+
             int rowAffected = await Task.Run(() => repository.DeletePerson(id));
 
             if (rowAffected > 0) return Ok();
 
-            return BadRequest();
+            return NoContent();
         }
     }
 }
